requirements.txt
fastapi
uvicorn[standard]
pydantic
pytest
httpx
sqlalchemy
passlib[bcrypt]
pyjwt
alembic
pydantic-settings
python-dotenv


